LayoutTests/crypto/webkitSubtle/hmac-sign-verify-expected.txt

CONSOLE MESSAGE: line 20: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
Test HMAC sign and verify functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw HMAC key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 8
PASS key.algorithm.hash.name is 'SHA-1'
PASS key.usages is ['sign', 'verify']
Using the key to sign 'foo'...
PASS bytesToHexString(new Uint8Array(signature)) is 'bebbc02e46b0f81183f40c25dce23e5045d65519'
Verifying the signature...
PASS verificationResult is true
PASS successfullyParsed is true

TEST COMPLETE
